btb/pyproject.toml
[project]
name = "bot-the-builer-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "chromadb>=0.6.3",
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "grpcio>=1.70.0",
    "mdextractor>=0.0.3",
    "numpy>=2.2.3",
    "openai>=1.65.2",
    "psycopg2-binary>=2.9.10",
    "weave>=0.51.35",
    "weaviate>=0.1.2",
]
